Implemented "mixRecorder" class
- draws a circular REC button (w/ label on the right side of the button)
- records audio from the master node
- saves file to disk

NOTE: in SAFARI web browser the ‘save file’ action will leave
index.html page forcing the user to save the page source as an audio
file. This is far from ideal but I’m not sure how to solve this right
now. In MOZILLA a window pops up asking the user to save the file to
disk —> this would be the desired method for all browsers….